+++ Ya/Program/Interface.hs
@@ -90,3 +90,24 @@ instance Covariant Endo Semi Functor (->) t
 		\(Construct x old) -> These
 			(T_TT_I / wrap @Arrow @(R_U_I_T_I _ _ _) `fo` old)
 			(\new -> Construct x / unwrap @Arrow @(R_U_I_T_I _ _ _) `fo` unwrap new)
+
+type family Ramification datastructure where
+	Ramification (Tree (U_I_I (/\) `T_TT_I` Optional)) = Unit \/ Unit
+
+type family Branching datastructure where
+	Branching (Tree (U_I_I (/\) `T_TT_I` Optional)) = Optional `T_TT_I` Binary Tree
+
+class Hierarchial datastructure => Brancheable datastructure where
+	branch :: Ramification datastructure ->
+		Attribute (datastructure item) (Branching datastructure item)
+
+-- TODO: refactor using limits
+instance Brancheable (Tree (U_I_I (/\) `T_TT_I` Optional)) where
+	branch p = W_I_II_II `compose` U_I_UU_III_U_II_I /
+		\(Construct x (T_TT_I (U_I_I (These lb rb)))) -> These
+			/ case p of
+				This _ -> T_TT_I @Optional (wrap @(->) `fo` lb)
+				That _ -> T_TT_I @Optional (wrap @(->) `fo` rb)
+			/ \new -> case p of
+				This _ -> Construct x (T_TT_I (U_I_I (These (unwrap @(->) `fo` unwrap new) rb)))
+				That _ -> Construct x (T_TT_I (U_I_I (These lb (unwrap @(->) `fo` unwrap new))))